FOLDER REORGANIZATION
Moved everything into model, view, and controller packages to make things a little easier and more MVC-oriented.
