+++ src/controller/GUIController.java
@@ -1,0 +1,19 @@
+package controller;
+
+public class GUIController {
+
+    // === VARIABLES AND FIELDS ===
+    private boolean inBattleScene, inStoryScene;
+
+
+    // === GUI CONTROLLER CONSTRUCTOR ===
+    public GUIController(){
+        this.inBattleScene = false;
+        this.inStoryScene = false;
+    }
+
+
+
+
+
+}
