--- src/ety/BattleChoice.java
@@ -1,14 +1,0 @@
-package ety;
-
-public enum BattleChoice {
-
-    // === ENUM VALUES ===
-    // -- Battle choices --
-    ATTACK,
-    DEFEND,
-    USE_ITEM,
-    RUN,
-
-    // -- default choice --
-    INVALID
-}
