--- src/gui/controllers/GUIController.java
@@ -1,19 +1,0 @@
-package gui.controllers;
-
-public class GUIController {
-
-    // === VARIABLES AND FIELDS ===
-    private boolean inBattleScene, inStoryScene;
-
-
-    // === GUI CONTROLLER CONSTRUCTOR ===
-    public GUIController(){
-        this.inBattleScene = false;
-        this.inStoryScene = false;
-    }
-
-
-
-
-
-}
