--- src/gui/parts/ButtonPanel.java
@@ -1,25 +1,0 @@
-package gui.parts;
-
-import javax.swing.*;
-import java.awt.*;
-import java.util.ArrayList;
-
-public class ButtonPanel extends JPanel {
-
-    // === VARIABLES AND FIELDS ===
-    private final FlowLayout panelLayout;
-
-
-    // === BUTTON PANEL CONSTRUCTOR ===
-    public ButtonPanel(){
-        this.panelLayout = new FlowLayout();
-        this.setLayout(panelLayout);
-    }
-
-
-
-
-
-
-
-}
