--- src/gui/parts/buttons/GameButtons.java
@@ -1,4 +1,0 @@
-package gui.parts.buttons;
-
-public enum GameButtons {
-}
