+++ src/model/ety/BattleChoice.java
@@ -1,0 +1,14 @@
+package model.ety;
+
+public enum BattleChoice {
+
+    // === ENUM VALUES ===
+    // -- Battle choices --
+    ATTACK,
+    DEFEND,
+    USE_ITEM,
+    RUN,
+
+    // -- default choice --
+    INVALID
+}
