+++ src/model/ety/enemy/Slime.java
@@ -1,0 +1,21 @@
+package model.ety.enemy;
+
+import model.ety.StatBlock;
+
+public class Slime extends Enemy{
+
+    // === CONSTRUCTOR FOR SLIME ===
+    public Slime(int level){
+        super(
+                "Slime",
+                "A gooey, disgusting blob!",
+                new StatBlock(
+                        level,
+                        level + 9,
+                        level,
+                        level - 1,
+                        level - 1
+                ));
+    }
+
+}
