+++ src/view/parts/ButtonPanel.java
@@ -1,0 +1,24 @@
+package view.parts;
+
+import javax.swing.*;
+import java.awt.*;
+
+public class ButtonPanel extends JPanel {
+
+    // === VARIABLES AND FIELDS ===
+    private final FlowLayout panelLayout;
+
+
+    // === BUTTON PANEL CONSTRUCTOR ===
+    public ButtonPanel(){
+        this.panelLayout = new FlowLayout();
+        this.setLayout(panelLayout);
+    }
+
+
+
+
+
+
+
+}
