+++ src/view/parts/buttons/GameButtons.java
@@ -1,0 +1,4 @@
+package view.parts.buttons;
+
+public enum GameButtons {
+}
